UWMail (Alpha 1)
This project is an email server that reads database info from a zip file
on startup.
Users can view achieved conversations and scroll through emails in a
command line interface.
Users are also able to delete emails and view the trash.
All emails are identified by a unique message id, timestamp and other
relevant info related to the conversation.